Rename daybreak persistence strategy
