+++ lib/mumukit/auth.rb
@@ -13,7 +13,7 @@ require_relative './auth/token'
 require_relative './auth/scope'
 require_relative './auth/permissions'
 require_relative './auth/store'
-require_relative './auth/permissions_persistence'
+require_relative './auth/permissions_persistence/daybreak'
 
 require 'ostruct'
